+++ multiqc/modules/bustools/bustools.py
@@ -44,6 +44,10 @@ class MultiqcModule(BaseMultiqcModule):
             self.bustools_data[s_name] = content
             self.add_data_source(f)
 
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, s_name)
+
         # Filter to strip out ignored sample names
         self.bustools_data = self.ignore_samples(self.bustools_data)