+++ multiqc/modules/filtlong/filtlong.py
@@ -25,6 +25,10 @@ class MultiqcModule(BaseMultiqcModule):
         for f in self.find_log_files("filtlong", filehandles=True):
             self.parse_logs(f)
 
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, f["s_name"])
+
         self.filtlong_data = self.ignore_samples(self.filtlong_data)
 
         if len(self.filtlong_data) == 0: