+++ multiqc/modules/mapdamage/mapdamage.py
@@ -35,6 +35,10 @@ class MultiqcModule(BaseMultiqcModule):
         for f in self.find_log_files("mapdamage", filehandles=True):
             self.parse_logs(f)
 
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, f["s_name"])
+
         # Filter to strip out ignored sample names
         self.threepGtoAfreq_data = self.ignore_samples(self.threepGtoAfreq_data)
         self.fivepCtoTfreq_data = self.ignore_samples(self.fivepCtoTfreq_data)