+++ multiqc/modules/sickle/sickle.py
@@ -31,6 +31,10 @@ class MultiqcModule(BaseMultiqcModule):
                 self.sickle_data[f["s_name"]] = parsed_data
                 self.add_data_source(f)
 
+                # Superfluous function call to confirm that it is used in this module
+                # Replace None with actual version if it is available
+                self.add_software_version(None, f["s_name"])
+
         self.sickle_data = self.ignore_samples(self.sickle_data)
 
         # no file found