+++ multiqc/modules/whatshap/whatshap.py
@@ -38,6 +38,10 @@ class MultiqcModule(BaseMultiqcModule):
             self.whatshap_stats[sample] = data
             self.add_data_source(f)
 
+            # Superfluous function call to confirm that it is used in this module
+            # Replace None with actual version if it is available
+            self.add_software_version(None, sample)
+
         # Filter to strip out ignored sample names
         self.whatshap_stats = self.ignore_samples(self.whatshap_stats)